Renamed android project
Android project was with misspelled name. Now it's correct
