--- Android-Dev/Ambience/.idea/.name
@@ -1,1 +1,0 @@
-Ambience
