--- Android-Dev/Ambience/app/src/main/java/com/gbotelhob/ambience/MainActivity.java
@@ -1,12 +1,0 @@
-package com.gbotelhob.ambience;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-}
