+++ Android-Dev/Ambiencer/.gitignore
