+++ Android-Dev/Ambiencer/app/.gitignore
