+++ Android-Dev/Ambiencer/app/src/main/java/com/gbotelhob/ambiencer/MainActivity.java
@@ -1,0 +1,12 @@
+package com.gbotelhob.ambiencer;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
