+++ Android-Dev/Ambiencer/app/src/main/res/layout/scene_list_item.xml
