+++ Android-Dev/Ambiencer/app/src/main/res/values/strings.xml
