+++ Android-Dev/Ambiencer/gradle.properties
